+++ src/culsans/_proxies.py
@@ -69,8 +69,8 @@ class SyncQueueProxy(SyncQueue[_T]):
     def join(self) -> None:
         self.wrapped.sync_join()
 
-    def task_done(self) -> None:
-        self.wrapped.task_done()
+    def task_done(self, count: int = 1) -> None:
+        self.wrapped.task_done(count)
 
     def shutdown(self, immediate: bool = False) -> None:
         self.wrapped.shutdown(immediate)
@@ -151,8 +151,8 @@ class AsyncQueueProxy(AsyncQueue[_T]):
     async def join(self) -> None:
         await self.wrapped.async_join()
 
-    def task_done(self) -> None:
-        self.wrapped.task_done()
+    def task_done(self, count: int = 1) -> None:
+        self.wrapped.task_done(count)
 
     def shutdown(self, immediate: bool = False) -> None:
         self.wrapped.shutdown(immediate)