+++ CHANGELOG.md
@@ -16,6 +16,18 @@ and this project adheres to
 Commit messages are consistent with
 [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
+[Unreleased]
+------------
+
+### Added
+
+- `count` parameter to the `task_done()` methods to identify that the specified
+  number of tasks has been completed. Useful for subclasses that implement
+  flattened queues.
+- `_isize()` and `_chain()` overridable methods, which allow to create
+  subclasses that implement sequence/flattened queues. Solves
+  [#9](https://github.com/x42005e1f/culsans/issues/9).
+
 [0.10.0] - 2025-11-04
 ---------------------
 
@@ -223,6 +235,7 @@ Commit messages are consistent with
   than `janus.Queue` in multi-threaded tests, simplifies usage, and expands the
   number of supported use cases (multiple event loops, `trio` support, etc.).
 
+[unreleased]: https://github.com/x42005e1f/culsans/compare/0.10.0...HEAD
 [0.10.0]: https://github.com/x42005e1f/culsans/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/x42005e1f/culsans/compare/0.8.0...0.9.0
 [0.8.0]: https://github.com/x42005e1f/culsans/compare/0.7.1...0.8.0